front: decouple DFM trigger
